feat!: migrate to cent-based money pattern and introduce RSV licensing model
Implementiert das "Money Pattern" (Fowler) über den gesamten Stack hinweg.
Alle Geldbeträge werden nun strikt als ganzzahlige Cents (`INTEGER`)
in der Datenbank gespeichert und über die API kommuniziert, um
Rundungsfehler bei Floating-Point-Arithmetik zu vermeiden. Die DB-Migration
konvertiert bestehende Datensätze automatisch.

Gleichzeitig wird das statische Lizenz-Modell (Web/Print/Original) durch
einen dynamischen Lizenz-Katalog (Use Cases & Modifikatoren) nach
RSV-Standard ersetzt. Ein neues Management-Interface inkl. `super_admin`
Middleware wurde hinzugefügt, um diese Kataloge zu pflegen.

BREAKING CHANGE: API-Contracts für Preise, Warenkörbe, Bestellungen und
Produkte erwarten und liefern ab sofort Cents statt Euro-Dezimalwerte.
Das CartItem-Interface im Frontend wurde stark modifiziert und ist nicht
mehr abwärtskompatibel zu alten lokalen Warenkörben.

+++ backend/app/Models/Product.php
@@ -9,5 +9,5 @@ class Product extends Model
 {
     use HasUuids;
     protected $fillable = ['type', 'name', 'description', 'price'];
-    protected $casts = ['price' => 'float'];
+    protected $casts = ['price' => 'integer'];
 }
